Feat(project 01): add a simple project
Description:
In this commit, I have added a simple test project to the main project.
The concept of this project is to increase the level of familiarity and
awareness of how the methods of sending and receiving data in gRPC work.
In this project, the implementation and operation of unary, server stream,
client stream and bidi stream operations have been investigated.

Change Logs:
[Echo proto file] [Added] - Add a echo system protocol buffer.
[Server file] [Added] - Add and initialize a gRPC server.
[Functions file] [Added] - Add gRPC server request manager functions.
[Client file] [Added] - Add a client file to send and recieve data to server.